% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AOCF}
\alias{AOCF}
\title{Average Observation Carried Forward}
\usage{
AOCF(x)
}
\arguments{
\item{x}{data series/vector with missing values to fill}
}
\value{
returns the same data series with missing values filled in with the average numeric value rounded to 2 decimal places
}
\description{
This a function to handle missing values in a dataframe column. It fills empty/missing rows values
with the average numeric value.
}
\examples{
AOCF(c(1, 2, NA, 5, NA, 9, 10))
}
