% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASL}
\alias{ASL}
\title{Dummy ASL Dataset for Package Usage Examples}
\format{A data frame based off a diabetic foot ulcer study with 60 rows and 14 variables:
\describe{
  \item{USUBJID}{unique subject identifier}
  \item{ARM}{treatment arm}
  \item{TRT}{active vs placebo label}
  \item{CHRT}{treatment arms of interest grouping}
  \item{CHRTGRP}{pooled treatment arms}
  \item{BAGE}{baseline age}
  \item{AGE65}{categorical age below 65 vs. not}
  \item{REGION}{study region}
  \item{BASEUA}{cutoff category for baseline ulcer area}
  \item{BASE}{baseline ulcer area in cm^2}
  \item{WEIGHT}{baseline weight in kg}
  \item{HBA1C}{baseline HBA1C count}
  \item{SAFFL}{safety evaluable flag}
  \item{ITTFL}{intent-to-treat flag}
  ...
}}
\usage{
ASL
}
\description{
Dummy Dataset of Analysis Dataset for Subject Level records
}
\keyword{datasets}
