% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATE}
\alias{ATE}
\title{Dummy ATE Dataset for Package Usage Examples}
\format{A data frame based off a diabetic foot ulcer study with 120 rows and 19 variables:
\describe{
  \item{USUBJID}{unique subject identifier}
  \item{PARAMCD}{parameter code for time to event}
  \item{PARAM}{full description of time to event parameter}
  \item{CNSR}{censorship variable}
  \item{AVAL}{analysis value}
  \item{AVALU}{analysis value unit}
  \item{ARM}{treatment arm}
  \item{TRT}{active vs placebo label}
  \item{CHRT}{treatment arms of interest grouping}
  \item{CHRTGRP}{pooled treatment arms}
  \item{BAGE}{baseline age}
  \item{AGE65}{categorical age below 65 vs. not}
  \item{REGION}{study region}
  \item{BASEUA}{cutoff category for baseline ulcer area}
  \item{BASE}{baseline ulcer area in cm^2}
  \item{WEIGHT}{baseline weight in kg}
  \item{HBA1C}{baseline HBA1C count}
  \item{SAFFL}{safety evaluable flag}
  \item{ITTFL}{intent-to-treat flag}
  ...
}}
\usage{
ATE
}
\description{
Dummy Dataset of Analysis Dataset for Time to Event records
}
\keyword{datasets}
