% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BM}
\alias{BM}
\title{Dummy BM Dataset for Package Usage Examples}
\format{A data frame based off a diabetic foot ulcer study with 720 rows and 10 variables:
\describe{
  \item{USUBJID}{unique subject identifier}
  \item{CHRT}{treatment arms of interest grouping}
  \item{CHRTGRP}{pooled treatment arms}
  \item{BASEUA}{cutoff category for baseline ulcer area}
  \item{ANALYTE}{name of biomarker analyte}
  \item{ADY}{analysis (study) day}
  \item{AVAL}{analysis value (biomarker lab value)}
  \item{BASE}{baseline ulcer area in cm^2}
  \item{CHG}{change in biomarker from baseline}
  \item{PCHG}{percent change in biomarker from baseline}
  ...
}}
\usage{
BM
}
\description{
Dummy Dataset of Analysis Dataset for biomarker records
}
\keyword{datasets}
