% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Box Plot}
\usage{
boxplot(data, group_by, y, color_by, facet = NULL, subj = "USUBJID",
  title = "", xlab = "Visit Day", ylab = "AVAL", caption = "")
}
\arguments{
\item{data}{a dataframe containing the input data to plot}

\item{group_by}{x-axis grouping variable}

\item{y}{column name for y-axis parameter}

\item{color_by}{variable to color by}

\item{facet}{optional variable to facet plot by}

\item{subj}{column name for unique subject identifier, default is "USUBJID"}

\item{title}{custom plot title, passed as a string}

\item{xlab}{custom x-axis plot label, passed as string}

\item{ylab}{custom y-axis plot label, passed as a string}

\item{caption}{optional string to include as a caption on bottom of plot}
}
\value{
produces a box plot with optional facet parameters
}
\description{
Generate a flexible box plot for endpoint analysis
}
\examples{
boxplot(data = AZA, group_by = "REGION", y = "PDCR", color_by = "BASEUA", facet = "TRT")
}
