% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdplot.R
\name{grdplot}
\alias{grdplot}
\title{Gradient Plot}
\usage{
grdplot(type = "heat", data, x, y, facet = NULL, ord,
  reverse = FALSE, ncol = 3, subj = "USUBJID", title = "",
  xlab = "Subjects", ylab = "Study Day", leglab = "", caption = "")
}
\arguments{
\item{type}{use either "heat" or "bubble", controls the type of plot produced (string)}

\item{data}{a dataframe containing the input data to plot (dataframe)}

\item{x}{column name for x-axis parameter (string)}

\item{y}{column name for y-axis parameter (string)}

\item{facet}{optional parameter, column name containing a subgrouping factor to facet the plot by (string)}

\item{ord}{column name for parameter to order color gradient by (string)}

\item{reverse}{optional boolean perameter for flipping the size of the bubbles
and for "negative" measurements like decrease/percent decrease (logic)}

\item{ncol}{optional parameter to control number of facets to display
in a row if facet variable is provided, must be at least 1, default value is 3 (integer)}

\item{subj}{column name for unique subject identifier used to compute counts in legend, default is "USUBJID" (string)}

\item{title}{optional parameter, custom plot title (string)}

\item{xlab}{optional parameter, custom x-axis plot label (string)}

\item{ylab}{optional parameter, custom y-axis plot label (string)}

\item{leglab}{descriptive label for heat map legend (string)}

\item{caption}{optional string to include as a caption on bottom of plot (string)}
}
\value{
produces a gradient plot with optional facet parameters
}
\description{
Generate a flexible gradient plot (heatmap or bubble plot) for efficacy analysis
}
\examples{
grdplot(
  type = "bubble", data = AZA, x = "RANK", y = "ADY", facet = "CHRTGRP", ord = "PDCR",
  reverse=TRUE, ncol = 2, xlab = "Subject ID", leglab = "\% Decrease in UA"
)

grdplot(
  type = "heat", data = AZA, x = "RANK", y = "ADY", facet = "CHRTGRP", ord = "PDCR",
  reverse=TRUE, ncol = 2, xlab = "Subject ID", leglab = "\% Decrease in UA"
)

grdplot(
  type = "bubble", data = AZA, x = "RANK", y = "ADY", facet = "CHRTGRP", ord = "AVAL",
  ncol = 2, xlab = "Subject ID", leglab = "Ulcer Area"
)
}
