% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LOCF}
\alias{LOCF}
\title{Last Observation Carried Forward}
\usage{
LOCF(x)
}
\arguments{
\item{x}{data series/vector with missing values to fill (vector)}
}
\value{
returns the same data series with missing values filled in with the preceding value
}
\description{
This a function to handle missing values in a dataframe column. It fills empty/missing rows values
with the immediate available previous value.
}
\examples{
LOCF(c(1, 2, NA, 5, NA, 9, 10))
}
