% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longplot.R
\name{longplot}
\alias{longplot}
\title{Longitudinal Plot}
\usage{
longplot(data, group_by, x, y, subj = "USUBJID", stat = NULL,
  markers = NULL, title = "", xlab = "Days", ylab = "AVAL",
  caption = "")
}
\arguments{
\item{data}{a dataframe containing the input data to plot}

\item{group_by}{a column name containing factors to split the plot by}

\item{x}{column name for x-axis parameter}

\item{y}{column name for y-axis parameter}

\item{subj}{column name for unique subject identifier, default is "USUBJID"}

\item{stat}{type of summary statistic to compute on y-variable, choose "mean" or "median" if y is not already summary-level data}

\item{markers}{optional parameter, list of special interest x-axis values to display vertical dashed marker lines}

\item{title}{custom plot title, passed as a string}

\item{xlab}{custom x-axis plot label, passed as string}

\item{ylab}{custom y-axis plot label, passed as a string}

\item{caption}{optional string to include as a caption on bottom of plot}
}
\value{
produces a line plot with study endpoint measurements over time
}
\description{
Generate a flexible longitudinal plot for endpoint analysis
}
\examples{
longplot(data = AZA, group_by = "CHRT", x = "ADY", y = "AVAL", stat = "mean", markers = c(45,85))
}
