% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round}
\alias{round}
\title{Round}
\usage{
round(x, n)
}
\arguments{
\item{x}{the number to round}

\item{n}{the number of decimal places to round}
}
\value{
returns a rounded version of the value provided
}
\description{
This is a custom round function that rounds a number to the nearest whole number down
if the decimal is below 0.5 and up if it is above. Default R round behavior rounds decimals of 0.5 down
when the whole number is even and this function avoids that behavior.
}
\examples{
round(4.5, 0)
}
