% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgtplot.R
\name{spgtplot}
\alias{spgtplot}
\title{Spaghetti Plot}
\usage{
spgtplot(data, group_by, x, y, subj = "USUBJID", title = "",
  xlab = "Visit Day", ylab = "AVAL", symb = NULL, caption = "")
}
\arguments{
\item{data}{a dataframe containing the input data to plot}

\item{group_by}{a column name containing factors to split the plot by}

\item{x}{column name for x-axis parameter}

\item{y}{column name for y-axis parameter}

\item{subj}{column name for unique subject identifier, default is "USUBJID"}

\item{title}{custom plot title, passed as a string}

\item{xlab}{custom x-axis plot label, passed as string}

\item{ylab}{custom y-axis plot label, passed as a string}

\item{symb}{optional parameter to shape markers by a factor type column}

\item{caption}{optional string to include as a caption on bottom of plot}
}
\value{
produces a spaghetti plot with study endpoint measurements over time that can be facet by a factor-class column variable
}
\description{
Generate a flexible facet spaghetti plot for endpoint analysis
}
\examples{
spgtplot(data = AZA, group_by = "CHRTGRP", x = "ADY", y = "AVAL", symb = "EPOCH")
}
