% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strToFileName}
\alias{strToFileName}
\title{String to File Name}
\usage{
strToFileName(s, sep = "-")
}
\arguments{
\item{s}{string parameter to convert to (string)}

\item{sep}{optional parameter, symbol to separate words by, default is "-" (string)}
}
\value{
returns a parsed string formatted better for filenames
}
\description{
This function parses a string to a format easier to use in filenames by removing spaces and replacing
with a symbol like "-" and converting all letters to lowercase and removing special symbols.
}
\examples{
strToFileName(s = "This is an example FiLeNme with# (symbols >=4) to remove")
}
