% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tteplot.R
\name{tteplot}
\alias{tteplot}
\title{Time to Event Plot}
\usage{
tteplot(data, group_by, param, paramcd = "PARAMCD", aval = "AVAL",
  cnsr = "CNSR", reverse = FALSE, confint = 0.8, title = "",
  xlab = "", ylab = "")
}
\arguments{
\item{data}{a dataframe containing the input data to plot}

\item{group_by}{a column name containing factors to split the plot by}

\item{param}{the time-to-event parameter to plot (e.g. "OS")}

\item{paramcd}{the column name containing the parameter, default is "PARAMCD"}

\item{aval}{the column name containing the numeric analysis value, default is "AVAL"}

\item{cnsr}{the column name containing the censor value, default is "CNSR"}

\item{reverse}{boolean value controls whether the survival formula should be flipped, default is FALSE: Surv(CNSR - 1), if set to TRUE then behaves like Surv(1 - CNSR)}

\item{confint}{confidence interval to calculate with, default is 0.8}

\item{title}{custom plot title, passed as a string}

\item{xlab}{custom x-axis plot label, passed as string}

\item{ylab}{custom y-axis plot label, passed as a string}
}
\value{
produces a Kaplan-Meier type survival plot
}
\description{
Generate a flexible KM plot with a risk table
}
\examples{
tteplot(data = ATE, group_by = "CHRT", param = "TFUC", reverse = TRUE)
}
