% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WOCF}
\alias{WOCF}
\title{Worst Observation Carried Forward}
\usage{
WOCF(x)
}
\arguments{
\item{x}{data series/vector with missing values to fill}
}
\value{
returns the same data series with missing values filled in with the lowest alpha/numeric value
}
\description{
This a function to handle missing values in a dataframe column. It fills empty/missing rows values
with the worst(lowest) available alpha/numeric value.
}
\examples{
WOCF(c(1, 2, NA, 5, NA, 9, 10))
}
