% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AOCF}
\alias{AOCF}
\title{Average Observation Carried Forward}
\usage{
AOCF(x)
}
\arguments{
\item{x}{data series/vector with missing values to fill}
}
\value{
returns the same data series with missing values filled in with the average numeric value
}
\description{
This a function to handle missing values in a numeric dataframe column. It fills empty/missing rows values
with the average numeric value rounded to the minimum number of decimal places in the input data.
}
\examples{
AOCF(c(NA, 2.230, 1.1, 2390.141, NA, 1341.012, 10.24))
}
