% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Box Plot}
\usage{
boxplot(data, group_by, y, color_by, facet = NULL, subj = "USUBJID",
  title = "", xlab = "Visit Day", ylab = "AVAL", caption = "")
}
\arguments{
\item{data}{a dataframe containing the input data to plot (dataframe)}

\item{group_by}{x-axis grouping variable (string)}

\item{y}{column name for y-axis parameter (string)}

\item{color_by}{variable to color by (string)}

\item{facet}{optional parameter, column name containing a subgrouping factor to facet the plot by (string)}

\item{subj}{column name for unique subject identifier used to compute counts in legend, default is "USUBJID" (string)}

\item{title}{optional parameter, custom plot title (string)}

\item{xlab}{optional parameter, custom x-axis plot label (string)}

\item{ylab}{optional parameter, custom y-axis plot label (string)}

\item{caption}{optional string to include as a caption on bottom of plot (string)}
}
\value{
produces a box plot with optional facet parameters
}
\description{
Generate a flexible box plot for endpoint analysis
}
\examples{
boxplot(
  data = AZA, group_by = "REGION", y = "PDCR", color_by = "BASEUA", facet = "TRT",
  title = "Comparison of Treatment", xlab = "Region", ylab = "\% Decrease in Ulcer Area",
  caption = "Colored by baseline ulcer area"
)
}
