% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdplot.R
\name{grdplot}
\alias{grdplot}
\title{Gradient Plot}
\usage{
grdplot(type = "heat", data, x, y, facet = NULL, ord, ncol = 3,
  subj = "USUBJID", title = "", xlab = "Subjects", ylab = "Study Day",
  leglab = "", caption = "")
}
\arguments{
\item{type}{controls the type of plot produced}

\item{data}{a dataframe containing the input data to plot}

\item{x}{column name for x-axis parameter}

\item{y}{column name for y-axis parameter}

\item{facet}{optional variable to facet plot by}

\item{ord}{column name for parameter to order color gradient by}

\item{ncol}{optional parameter to control number of facets to display
in a row if facet variable is provided, default is 3}

\item{subj}{column name for unique subject identifier, default is "USUBJID"}

\item{title}{custom plot title, passed as a string}

\item{xlab}{custom x-axis plot label, passed as string}

\item{ylab}{custom y-axis plot label, passed as a string}

\item{leglab}{descriptive label for heat map legend}

\item{caption}{optional string to include as a caption on bottom of plot}
}
\value{
produces a gradient plot with optional facet parameters
}
\description{
Generate a flexible gradient plot (heatmap or bubble plot) for efficacy analysis
}
\examples{
grdplot(type = "bubble", data = AZA, x = "RANK", y = "ADY", facet = "CHRT", ord = "PDCR")
}
